move App.js to components, add axios & Container
