+++ src/components/App.js
@@ -1,0 +1,40 @@
+import { useState } from 'react';
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+import { ThemeProvider } from 'styled-components';
+import { themes } from '../styles/themes.style';
+import { Container, Themebtn } from '../styles/styles.style';
+import GlobalStyle from '../styles/global.style';
+import Navbar from './Navbar';
+import Home from './Home';
+import Spirit from './Spirit';
+import Opportunity from './Opportunity';
+import Curiosity from './Curiosity';
+
+function App() {
+  const [currentTheme, setCurrentTheme] = useState('light');
+  const getOppositeTheme = () => ((currentTheme === 'light') ? 'dark' : 'light');
+
+  return (
+    <ThemeProvider theme={themes[currentTheme]}>
+      <GlobalStyle />
+      <Router>
+        <Navbar />
+        <Themebtn onClick={() => setCurrentTheme(getOppositeTheme())}>
+          <h1>
+            {(currentTheme === 'light') ? <i className="fas fa-moon" /> : <i className="fas fa-sun" />}
+          </h1>
+        </Themebtn>
+        <Container>
+          <Routes>
+            <Route exact path="/" element={<Home />} />
+            <Route path="/curiosity" element={<Curiosity />} />
+            <Route path="/opportunity" element={<Opportunity />} />
+            <Route path="/spirit" element={<Spirit />} />
+          </Routes>
+        </Container>
+      </Router>
+    </ThemeProvider>
+  );
+}
+
+export default App;
