+++ src/styles/styles.style.js
@@ -1,9 +1,17 @@
 import styled from 'styled-components';
 
+const Container = styled.div`
+    height: 100vh;
+    width: 100vw;
+    border: 5px solid black;
+    background-color: ${(props) => props.theme.backgroundColor};
+    color: ${(props) => props.theme.color};
+`;
+
 const Themebtn = styled.button`
     border: none;
     align-self: flex-start;
-    background-color: ${(props) => props.theme.backgroundColor};;
+    background-color: ${(props) => props.theme.button.backgroundColor};;
     color: #F7D047;
     border-radius: 8px;
     box-sizing: border-box;
@@ -16,5 +24,6 @@ const Themebtn = styled.button`
 `;
 
 export {
+  Container,
   Themebtn,
 };
