+++ src/styles/themes.style.js
@@ -1,11 +1,18 @@
+/* eslint-disable import/prefer-default-export */
 const lightTheme = {
   color: '#161617',
   backgroundColor: '#F3F3F3',
+  button: {
+    backgroundColor: '#181828',
+  },
 };
 
 const darkTheme = {
   color: '#F1F1F1',
   backgroundColor: '#181828',
+  button: {
+    backgroundColor: '#FFFFFF',
+  },
 };
 
 export const themes = {
